+++ contsub/exceptions.py
@@ -1,0 +1,17 @@
+class alreadyOpen(Exception):
+    pass
+
+class alreadyClosed(Exception):
+    pass
+
+class noBeamTable(Exception):
+    pass
+
+class tableExists(Exception):
+    pass
+
+class tableDimMismatch(Exception):
+    pass
+
+class CubeDimIsSmall(Exception):
+    pass
